Add `x` key to toggle showing/hiding a post.
I'm not sure this does anything beyond the current session.  That is,
reloading LI likely will show the post again.  If so, I may have to
switch to navigating the meatball menu instead.  And even then, it
seems like it takes a bit of time for the fact that posts were hidden
to percolate through LIs stack, so they are still likely to reshow.
And, if something else touches a post (e.g., a new comment, a new
friend likes the post, etc), it'll show back up.

Also, rearrange the code a tad bit so that keystroke registration is
at the top so it acts a bit like documentation.

Closes #3.