Change newPosts to loadNewPosts and update the keystroke from `n` to `l`.
This will also trigger the [Show more feed updates] button at the
bottom.

Normally updates are loaded automatically, but if the screen is
narrow, the [LinkedIn News] section is at the bottom instead and
loading does not happen.  This will now trigger whichever one is
active.  When the [New posts] button is active, the new messages will
show at the top.  Otherwise at the bottom.

Is also means that a user can preload as many messages as they want,
then scroll through leisurely.

A bit jittery as the focus will pop to the button at the bottom then
back to the top.

Closes #7.