Make sure that whatever the current element in focus counts as the current post (if applicable).
This works, mostly.  Is a little hinky.

If user navigates to a new post, hits `TAB` to another element in the
same post, then `j` to navigate down, it will actually refocus the
current post again.

Related, tabbing to an element outside of main, even it if meant going
through a bunch of posts, will keep the same old current post, rather
than the first or last.  Perhaps we need to do something like what is
done for `click` events instead.

Anyway, good enough for now.

Closes #12.