Support navigating through comments with `n` and `p`.
It is not perfect, but I learned some stuff along the way about
scrolling, and I'm going to revisit that as a whole.

Also leading into the situation where Posts exist in all sorts of
views.  The Feed, Groups, individual views, individual's recent
activity.  And so on.  So we want to learn how to handle all of those
in a similar way.  I sense a refactoring in my future.

Closes #21.