Manually highlight current post and comment.
Previously we were relying on the browser to highlight the current
element by forcing it into focus.  However, this is bad practice for
accessibility.  It also means that if focus moves away (say when
toggling a post, or click else where), the element would lose that
highlighting.  Also, it meant that only one of post or comment could
be highlighted at a time.  Now we can do both.

I'm not keen on the use of `!important` here, but LI has some other
CSS that was preventing comments from getting the desired effect.

Also, on post, the highlighted box is bigger than the focus ring was,
but I think it is all right.

Also, of course, the colors are icky, but those can be tuned later.

Issue #17.
Closes #9.