Handle the case where the element passed to `inInput` does not have a tagName.
Not sure if this is because of the previous change, or I just never
noticed this before.  Sometimes the element passed to isInput is
`document`, which does not have a tagName property.  And referring to
it would throw exceptions.  Really all this does is makes sure not to
clutter the log.