Switch to only tracking `focus` events for disabling hotkeys.
The previous implementation depended on catching `blur` events to
determine if a user left an input field so that hotkeys could be
reenabled.  However, it turns out that those events will not be fired
if the focused element is destroyed.  When a focused element is
destroyed, focus will change, causing another `focus` event, so that
can be used to see what really happened.

Did involved introducing a new global variable, but really, that is
what `setContext` is doing anyway, just hidden.

Closes #36.