Support the `/notifications/` page.
Adds navigating up and down posts, opening the menu, and activating
the notification (same as click on it).

Not perfect, but works well enough to ship and get feedback.

I had most of my notifications turned off, so I am unable to test
against most of them.  And it appears that almost every notification
has a different way of activating it.  So I'm sure I missed many.

Supporting a new page warrants a bump in the major version number.

Closes #35.