Do scroll to top of post if no comment is selected.
It turns out that if the user reads a lot of comments, then scrolls
back to the top of the post, the user cannot easily read the post
again.  We simply don't scroll in that situation.

We want to keep from scrolling in the situation where the users
requests more comments when in the middle of reading some.  But not
when they scroll off the top.

Issue #28.