Let users loop through loaded posts.
On initial load, no post is selected, and `j` would take the user to
the first post.  Now, in that state, `k` will take the user to the
last post.  While on the first or last post, they will go to the
opposite end (unlike comments, which will take the user back to the
containing post).  A little bit inconsisent, but I think appropriate.

One current problem is, when visiting the last post, it will often the
UI to load more posts.  This loading will interfere with keeping the
current post in view.  This problem exists else where as well, so a
common solution will be needed.  However, once posts are loaded,
navigating up or down then back will get to the correct post.

Issue #29.