Implement click-to-visit on notifications.
Not that it works all that well.  Clicking all notifications I
currently have is the same as activating it.  So by the time we
receive the click, all of the notifications have already been removed,
so unable to find it for tracking purposes.  When the user comes back
to the page, nothing can be found.

The only thing that keeps the notification around is clicking on the menu.

Closes #47.