Rewrite how we monitor for url changes.
Previously we were monitoring every single element added/removed to
the web page and using that as an execuse to check to see if the URL
changed.

Now we know that `div.authentication-outlet` is sufficient.  So use
`otmot` to monitor for that special div to show up, then register a
long lived `MutationObserver` to it.

I had hoped that we could monitor only that div's direct descendents
changing as the pages are being rewritten.  However, those particular
elements are removed and new ones added before the URL is updated (in
many cases).  As a result, we miss the change.

We could revisit and restrict this MO to the smaller subset, and if we
create a more thorough only until the URL changes.