Remove jitter from toggle+next action.
Use a ResizeObserver to detect when the toggled post collapses and
scroll back to the current post.

I'm not sure how this works so well.  I expected there to be more
jitter, but at least ending up at the right place.

I suspect that things just happened to get queued the right way so all
updates are resolved before the next screen refresh.  Pure luck, but
I'll take it.

There can still be a little bit of jitter as rendering of letters
seems to occasionally shift a pixel or two.  This may be related to
the scrollbar resizing when elements are reduced in size, and then
subsequent adjustements by the engine.  But they are very minor,
especially compared to the previous jitter.

Closes #17.