Remove jitter when explicitly loading more posts.
LinkedIn's UI will explicitly jump to the new posts at the end of the
page when loading.  Now detect when the containing div changes size
before scrolling back to our own idea of the current post.

Now the user can use `l` to load more posts while reading, and do so
multiple times while leisurely scrolling through currently loaded
posts.

Unfortunately this does not handle the case when going to the last
post with `>` and autoloading kicks in.  Since autoloading may or may
not kick in, handling that case may be different.