eslint: Investigate `no-ternary` and address a couple of issues.
I like the ternary operator, but agree it is not always the best
solution.  For instance, I just learned about the [Nullish coalescing
operator](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Nullish_coalescing)
and think that is a better solution in some of these case flagged.
But certainly not all.  And there are currently enough left over to
use selective disabling.  So, leaving it turned off for now.

A refactoring in progress will remove most of those, however, so when
that is done, I think it will be appropriate to revisit this.

Meanwhile, there are also a couple of similar ternary related rules
and while they do not trigger on any existing code, I went ahead and
enabled them.

Issue #68.