Factor out common navigation code to a new `Scroller` class.
And migrated comments to using it.

It should have most of the bits necessary to support are current
corner cases.  I'm sure that future ports will expose some that are
missing.

Also first foray into using JSDoc.  I'm sure I'm doing some of it
wrong.

The `_msg` method is probably over kill, but it sure was handy in
debugging.  Purposefully leaving it enabled for comments for the time
being.

This is a big enough refactor that I went ahead and bumped the middle
version number.

Closes #65.