Update initial debug message.
Now that we are using eslint on a regular basis, and that catches
parsing errors, we don't need the line to say anything about parsing.

However, now that it is at the end of the file, I have noticed it will
fail during initialization issues.

So might as well make it about that.