Save and restore current post AND comment before loading more.
This is not a perfect solution.  If the comment is not the bottom on
on the page, it will be.  So a little bit of jitter.  But at least in
the same realm.

Also, if the user does a lot of loads and navigating around through
comments at the same time, the UI may end up scrolling elsewhere (to
where new content is loading).  Either focus or another navigation
(comment or post), will bring it back into view.

Closes #79.