Add navigation through jobs within a jobs section.
Support next/previous with `n`/`p`.  Jobs can be activated (visited)
with the `Enter` key.

The first/last/focus actions (`<`, `>`, `f`) are context sensitive for
the section or job items.

Like for the Notifications page, the way to uniquely identify job
entries is complicated.  Lots of variations.

Known problem with losing track of the current job is more of anything
is loaded (sections or job items).

Closes #62.