Support some more notification types.
It appears that there are buttons as well as anchors with the
`.nt-card__headline` class.  Testing seems to indicate that only one
or the other appears.

It turns out that deleted notifications have neither.  But we can
detect the `Undo` button and click it.  If we ever see another `Apply
early` notification though, and it turns out it can be found with
`.nt-card__headline`, then we will probably drop `matchesKnownText()`,
and thus supporting the `Undo` through activation.

Issue #42.