Handle feed positioning better when loading more posts.
Introduced `otrot2`.  This function watches for resizes and simply
calls an action function every time it sees something.

Essentially, we are constantly trying to put the window back where it
was.

Works well for adding new posts at the end.  When the Load more posts
button comes up, however, just give up and jump to the first post.
Sorry.

Closes #96.