Make the `onClickSelector` a bit more precise.
Right now we call refresh at the some time we reenable the onClick
event handler.  No technical reason, was just convienent.

However, in this case, that calls refresh before enough of the
notifications view is rebuilt that the notification cards are not yet
present.

Here we try to make sure more of the page is repopulated before doing
the same stuff.

Most likely this is not sufficient, but perhaps *Good enough*.

Closes #103.