Configure a second MutationObserver for Jobs.
In order to capture changes to certain sections, we need to observe
changes to classes.  However, these need to be done for the subtree.
Monitoring the entire subtree for childList changes, however, is
excessive and results in both multiple matches and noticable
performance issues.

Since we are observing the same element
(div.scaffold-finite-scroll__content), we need two observers.
Fortunately we can reuse the same callback function.

Closes #86.