Migrate waiting for the navbar to appear to inside `LinkedIn`.
Upcoming changes require the navbar to already be present.  Since we
already have code waiting for it, makes sense to move that into this
class.

We then wait for that to finish before proceeding with the rest of the
initialization.  And this is probably a Good Thing (tm) anyway.

Issue #105.