eslint: Enable and configure `no-warning-comments`.
I believe that TODO and XXX are good in committed code as they point
to known weaknesses to be fix or code that needs special attention.  I
do wish eslint has a postfix decorator as well, so we could enforce a
pattern like `TODO(#\d+)`, but maybe there is a option I have not yet
seen.

Issue #68.