Make tabpanels have a scrollbar only when necessary.
With making nested items have scrollbars (such as the textarea in the
*Errors* panel), we end up with double scrollbars, one of which is
inactive.

So we either have the issue where some panels have scrollbars and
others do not, which makes text wrapping jump a bit (due the different
line lengths) as the user navigates through panels, or we end up with
double scrollbars.

I think dealing with the text wrapping changes is better than double
scrollbars.

Issue #94.