Move nav CSS styles back to the end of the list.
Apparently there are some LI provided styles that have the same
specificity for `section` elements, and so the LI ones, which were
later in the list, were being applied.  Essentially, no highlighting
for `section` elements.

This was only impacting the `/jobs/` page, so it went unnoticed for a
bit.

Issue #106.