Bring current section back into view if possible.
An interesting thing about the *My Network* view, is that the various
subsections change upon page reload.  If the user views a specific card, be it
user, group, whatever, then returns, that subsection may not be there upon the
reload.

For now, we give it 10 seconds to show up, then give up.  If it does show up,
we go back to it.  Otherwise, nothing happens.

Issue #142.

␄