Switch `SPA` over to using RegExp for matching pages.
This required getting rid of special handling for the global page.  As a
result, more than one page could be active at a time.  For instance, Global
matches all URLs, while a specific page may only match some URLs.

Now we have a set of active pages.  Moreover, the rewrite also handles the
case where, if there is a redirect, we end up with the exact same set of pages
matching, so nothing gets excessively disabled only to be immediately
reenabled.  Nicely, this fixes another bug!

Closes #146, #147.

␄