A fix to yesterday's *improvement* to CSS selectors.
My testing was not as careful as it should have been.

The *My Network* page has at least three different types of sections:
* Invitations (first section)
* More suggestions for you (last section)
* Everything else (in the middle)

Each type has a different way of selecting the items (mostly cards) inside of
them.  The tweak I made yesterday resulted in the *Everything else* type of
card to be selected twice.  Once as a 'li' element, and once as a 'section'
element.  And the uid calculation we use gave them the same id.  So the
`Scroller` would look for the next element, see it was id *X*, then end up
repicking the same element it was just on, stuck in an endless loop trying to
move forward.

Since my testing only involved move from the *See all* button to the first
card, nothing seemed amiss.  I will try to do better.

Issue #142.

␄