Wrap long lines in `TabbedUI`.
This contained some surprising policy decisions.

We had been trying to keep CSS consistent with:
`selector { prop: value; prop: value; }\n`

Now, being broken across lines of code, we still want to have some consistency
in how it is both presented in the JS, but also in the browser when debugging.

For the code, it goes somewhat like:
```
'selector {' +
  ' prop: value;' +
  ' prop: value;' +
  '}\n';
```
Which, in the browser, loses the last space before the closing `}`.
All-in-all, a minor annoyance, but still was something to think about.

Issue #148.

␄