Add `E`ngage as a way to interact with the current card.
I am not sold on this choice at all.  It is better than having separate keys
for each of Connect, Join, Follow, etc, but that verb... yuck.

I do not want to go with `ENTER` because, on this page, the actions are a bit
more aggressive than what is done on the *Notifications* page.  And we want to
keep these types of actions associated with upper-case letters.  So, subject
to change.

Issue #142.

␄