Group logging mode can now be explicitly set before hand.
It also affects any general logging statements while the group is active.  For
example, a group can be silenced before being seen, and not only will
intro/outro messages not be logged, neither will any messages in between.
Unless a new group is activated.

The API is still in flux.

Issue #145.

␄