Include `tagName` when watching mutations.
There are new job sections have have no identifying elements in them, so the
fallback was use the first line of innerText.  Unfortunately this meant the
section matched the first job entry in these new sections.  This confused some
other code, causing an infinite loop.

For this particular case, we can differentiate by including the element name
in the comparison.  Really, this is hack.  But since we hope to solve this
problem in a more general solution inside `Scroller` itself, this should not
be a problem.

Closes #155.

␄