Switch `Feed` to use the new `Scroller` click handler.
As a bonus feature, add click handling to the comments as well.  Kind of.  The
comment scroller has to exist before it can handle clicks and, currently, it
only exists after the first usage after a post is selected.  Need to figure
out the best approach for that before this is a complete example.

Issue #149.

␄