Enhance `Page` to take a `PageDetails` upon construction.
With #private variables, we need to move away from having code in the base
class that can use overrides in the subclasses.  Instead, we will need to pass
them in through the constructor or other mechanisms.  Here, we are adding some
items similar to what we already use, but can do differently.

Issue #130.

␄