Update focusing on the details pane.
The built in button does not do anything useful.  User would have to click
that button the TAB once.  So just go there directly.

Oddly enough, after selecting a new job card, hitting TAB TAB works as well,
so this is not much of a short cut, but it is already here.

Issue #143.

␄