Change hotkey from `Enter` to `c`.
In other cases where we can easily detect that focus is on an element inside
the current Scroller element.  There we can elect to allow the standard
browser function take over (e.g., TABbing to a link and hitting ENTER).

So here, we will use `n` and `p` to select a page, then `c` to change to it.

Issue #143.

␄