Change the job card `Scroller` to use a different element.
Like the recent change to the pagination one, use the surrounding `li`.  It
allows the highlights to shine through.

However, we are going to keep the auto-click on change feature.  Unlike
scrolling through the pages at the bottom, moving to a job card then selecting
it feels clunky.  Updating as we navigate feels natural.  Maybe it is due to
the dual pane layout.

Issue #143.

␄