Change the `pageReadySelector` for `JobCollections`.
It turns out that some view do not have enough job cards to require a results
paginator, in which case, waiting for one to show up will not work.

So, move to a different one.

However, since we were depending on the above to conviently make sure that it
was available to set the results page scroller, we had to change how that is
done as well.  The usual monitor stuff.

Amusingly enough the now breaks an invariant of `Logger`:  Groups will always
be nested.  Both 'activate' handlers gets invoked around the same time, so
both have active logging groups.  However, depending on how the page loads,
they may exit at different times.  Since we know what order these things will
finish, we could probably rearrange the order of the calls of `addService`,
but this is a good reminder that JS does have microtasks.

Issue #143.

␄