Initial implementation of saving `Logger` config.
I do not know if this applies to all userscript managers, but at least
Violentmonkey does not appear to use `JSON.stringify()` when saving values.
So we need cannot use the standard `toJSON()` method to convert.  Instead, we
will use `toPojo()`.

This is all disabled for now because I realized we need to first inform the
users that we will start saving data.

Issue #145.

␄