Implement `waitForSelector` and migrate users to it.
The most common use of `otmot` is waiting for elements to show up.  This
provides a simple way of doing that.

I realized this after looking at other JS libraries hosted on GreasyFork and
seeing how many implement this same pattern.  And that we will need it for
`Scroller`.

Issue #165.

␄