Migrate `Logger` to base.
Some cosmetic differences between the original version and this one:

* `GroupMode` is now internal.  I figured out how to use `static`
initialization blocks to create the known enums and freeze the nested class.
* `#Config` is now a private class.
* These classes required a lot of rearranging to meet the new style
guidelines.

I did waffle a bit about whether I should move `GroupMode` into `Logger` or
`Config` out.  With the way the library explicitly exposes symbols, both would
allow them to be effectively private.  Since this whole userscript thing is as
much of a learning exercise as anything else, I went with the hardway, simply
for the experience.  If `Logger` was going to be in a standalone module, I
would likely have moved them out, but since these libraries will not be that
granular, I also think this makes more sense.

Issue #167.

␄