Create `Logger.#Group` and move `GroupMode` into it.
We are soon going to have more items to track by group than just mode.

This does change the API a bit, but we had already make an incompatible change
when we made GroupMode private anyway.  Since we currently still only have one
known user, I think this is still a safe change to make.

Issue #145.

␄