Implement `+` and `-` for thumbs-up and thumbs-down on job cards.
This is a bit of a cheat.  Depending on the exact page being visited, the job
card may either have thumbs-{up,down} buttons OR a X dismiss button.

The way dismiss and thumbs-down are selected are identical (at the moment), so
dismiss just calls thumbs-down.

It is always possible that in the future, these two will have different ways
of selecting them and this will have to be changed again.

Closes #168.

␄