Replace a thenable with `await`.
Now that the wrapping IIFE is an `async` function, we can replace waiting on
`LinkedIn.ready`'s with an await rather than a `then()`.

This pausing the subsequent testing, which could get the `proceeding` log
message and those from creating `SPA` and the `Page` classes mixed in with it.

This does not eliminate all mixed in logs, since we do have event handling all
over the place.  But, it does reduce it.

Issue #167.

␄