Fix a bug in `Logger.#outro` and normalize `#intro`.
`#outro` was calling the actual log method after the group was popped, meaning
the message was being controlled by the wrong mode.  Since modes are fairly
new, this had gone unnoticed.

In fixing this, the actual logging was moved outside of the ending of the
group.  Updated `#intro` to be similar for consistency.  (Though both of these
will result in more accurate stats in the next change.)

Issue #145.

␄