Add a sequence number to `Logger`.
Each logger instance, and individual groups, are updated to the current
sequence upon use.  The current sequence number grows each time configurations
are restored from a saved state.

If exposed to the end users, they will be able to use the sequence number as
an indicator of the last time a particular logger or group was used.  Useful
if they want to clean up old settings they have (perhaps a logger was renamed,
or a particular group no longer exists).  It could also be used to rediscover
perhaps forgotten features of the program.

Issue #145.

␄