Tune `Jobs` pageReadySelector again.
It turns out that *#compactfooter-copyright* is loaded lazily.  So, if the
user scrolls down the Jobs page, visits a job, and comes back, the copyright
will not show up until until the user scrolls it into view.  So the page is
not activated until that happens.

However, the *aside* element, which will eventually contain the copyright, is
loaded earlier.  Still after *main*, though in that same batch.

Issue #165.

␄