Have `Scroller.#getItems` support *#containerItems*.
This may not be the final form, but it is simple.  Unlike how we
handled *#onClickElements*, where we changed a single HTML element to a set of
elements, here we are going from an (element, [string]) tuple to a list of
string, string pairs.  Unless we need to do massaging of the information while
poplating the list of elements to actually return, I suspect this may be the
simplest way to go.

Anyway, I suspect this may be the last thing necessary to support
containerItems.  We should start migrating the other major scrollers in a day
or two.

Issue #165.

␄