Move loading and saving of `Logger` configs outside of testing.
The lib/userscript variations will do sensible things when the {get,set}Value
functions are not granted, which is currently the case.  However, developers
can turn them on for testing.

This particular sequences, of saving on hidden and loading on visible, allows
developers to switch to another tab that has the userscript manager UI, tweak
the values there, and then load the new values when returning to the LinkedIn
tab.  At least, until we get an in-app UI built for this.

Issues #145, #170.

␄