Give the monitor function one last try in *otmot()*.
We already do similar in *otrot2()*, and the change there just makes the
coding look similar.

We also make use of this pattern in a recent change to `Scroller`.

Sometimes *otmot()* monitor functions do not inspect the records, but instead
it is easier to just poll a global state to see if it valid.  In those cases,
the *observeOptions* may not be precise enough, and an important mutation
missed.  This change gives everything a chance to resolve successfully.

A similar change was NOT made to *otrot()* because the options are set in the
function and anything missed is a bug in the library.

Issue #150.

␄