Add `Scroller` for `Messaging` view.
No hotkeys yet, but the basics are here, including finding the active card on
both page activate and changing the active *Focused*/*Other* tab.

When a new conversation is selected (and this happens without LIT), the focus
immediately moves to the *Write a message...* text area, which makes
navigation tricky.  The user would need to a nav-key TAB nav-key TAB sequence
to move around.

Need to review other pages to see if they have a similar problem and if we
already have precedence on how to handle this.

Issue #144.

␄