Move control of the tablist monitor into `Scroller` activation.
It turns out that having two services that control the same thing (current
item) starting up at the same time is a bit of a pain.  It works, but
functions are effectively multi-threading and really making the logger upset
that groups are exiting in the wrong order.  This is purely a cosmetic issue.

But, since it is easily fixed by this change, which also happens to reduce
code size a tiny bit, it is worth doing.

Issue #144.

␄