Switch monitoring of the message box to a `focus` event.
There are cases where the message box will get focus without setting the
`data-artdeco-is-focused` attribute.  This might be a bug in LinkedIn, but
does not matter.  It is not a good signal.

Also work around a quirk in `Scroller` that is needed for other containers
that constantly rebuild the elements being tracked.  This causes a 'change'
event for the old item then one for the new item.  And that would result in
nested event handlers.  It turns out if the user navigates at the right pace,
an outer handler could be removed, causing them all to be removed (same
callback function after all), and then leave the focus in the message box,
resulting in random navigation characters showing up.  One Page's fix is
another Page's problem.

Issues #144, #212.

␄