Implement `n`ext/`p`revious for entries inside a section.
Different sections have different layouts, so need a few different selectors
to figure them out.  Would not be surprised if a few are missed.

Sections that have tab panels inside of them (currently know about *Interests*
and *Recommendations*) are triggering a known issue with `Scroller` not
wrapping.  Fortunately, in practical use, not a big deal, however, it does
give us a good test bed to find a fix.

Issue #156, #208.

␄