Factor *roCallback()* out of *otrot()*.
This is to be consistent with *otmot()*.

While in here, also added an initial and final calls to the monitoring
function.  Previously the final call was explicitly avoided in *otrot()*, but
upon reflection, seems to cause no harm.  Similarly, the initial call was just
added to *otmot()* and seems reasonable here.

Issue #221.

␄