Call *getBoundingClientRect()* again after (maybe) scrolling.
It turns out that the result of *item.getBoundingClientRect()* is not a live
object, so after scrolling the bottom of the item into view, we need to
refetch the value to see if the top went out of view.

Adding this one line pushed the length of the method over the configured
limit, so also moved this *else* condition into a separate nested function.

Since doing that, also rid of the single property destructuring assignment
`const {item} = this;` and replaced it with a traditional assignment of
`const item = this.item`.

Closes #225.

␄