General improvements to use of RegExp in tests.
The recent experience with an overly forgiving expression suggests another
look at the whole set of tests, particularly the use of `.*`.

Hoisted the GUID expression to the file level and used in a few other
locations.

It is well known that using RegExp with HTML is generally a Bad Idea(tm).
These tests really should be rewritten to compare at the DOM level.

Issues #190, #254.

␄